backend/requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
google-generativeai
python-dotenv
pydantic